Move the Metrics resource interface next to the metric implementation.
