+++ pkg/reconciler/autoscaling/resources/metric.go
@@ -24,6 +24,21 @@ import (
 	"github.com/knative/serving/pkg/autoscaler"
 )
 
+// Metrics is an interface for notifying the presence or absence of metric collection.
+type Metrics interface {
+	// Get accesses the Metric resource for this key, returning any errors.
+	Get(ctx context.Context, namespace, name string) (*autoscaler.Metric, error)
+
+	// Create adds a Metric resource for a given key, returning any errors.
+	Create(ctx context.Context, metric *autoscaler.Metric) (*autoscaler.Metric, error)
+
+	// Delete removes the Metric resource for a given key, returning any errors.
+	Delete(ctx context.Context, namespace, name string) error
+
+	// Update update the Metric resource, return the new Metric or any errors.
+	Update(ctx context.Context, metric *autoscaler.Metric) (*autoscaler.Metric, error)
+}
+
 // MakeMetric constructs a Metric resource from a PodAutoscaler
 func MakeMetric(ctx context.Context, pa *v1alpha1.PodAutoscaler, config *autoscaler.Config) *autoscaler.Metric {
 	stableWindow, ok := pa.Window()
